DO NOT LOOK AT THIS CODE... SERIOUSLY..
