+++ component/index.js
@@ -3,6 +3,8 @@
 
 'use strict';
 
+require('./jquery-plugins');
+
 // Ouija component class
 var Ouija = require('./ouija');
 
@@ -16,21 +18,3 @@ var config = {
 var ouija = new Ouija(config);
 
 ouija.initialize();
-
-
-$.fn.serializeObject = function()
-{
-   var o = {};
-   var a = this.serializeArray();
-   $.each(a, function() {
-       if (o[this.name]) {
-           if (!o[this.name].push) {
-               o[this.name] = [o[this.name]];
-           }
-           o[this.name].push(this.value || '');
-       } else {
-           o[this.name] = this.value || '';
-       }
-   });
-   return o;
-};
