+++ component/templates/response.hbs
@@ -5,7 +5,7 @@
     <div class="ouija-author"><a href="#" alt="Timestamp">Jay Z</a></div>
     <div class="ouija-content">
         <form action="#">
-            <textarea name="content">The flow of the century, oh it's timeless, HOV!</textarea>
+            <textarea name="content" placeholder="Leave a comment..."></textarea>
             <footer>
                 <button class="text" class="ouija-response-cancel">Cancel</button>
                 <button class="ouija-response-save">Comment</button>
